requirements/base.txt
cssmin==0.2.0
Django==1.8.6
django-appconf==1.0.1
django-braces==1.8.1
django-compressor==1.5
django-crispy-forms==1.5.2
django-debug-toolbar==1.4
django-extensions==1.5.9
django-libsass==0.4
djangorestframework==3.3.1
docutils==0.12
factory-boy==2.6.0
fake-factory==0.5.3
futures==3.0.3
gunicorn==19.3.0
libsass==0.9.1
Markdown==2.6.4
Pillow==3.0.0
psycopg2==2.6.1
PyExecJS==1.1.0
Pygments==2.0.2
PyReact==0.5.2
pytz==2015.7
six==1.10.0
smartypants==1.8.6
sorl-thumbnail==12.3
sqlparse==0.1.18
typogrify==2.0.7
wheel==0.24.0
whitenoise==2.0.6
